Move ConfigProperties / ConfigurationException to SPI package and consistently pass ConfigProperties
