+++ extensions/aws/src/main/java/io/opentelemetry/extension/aws/AwsConfigurablePropagator.java
@@ -6,6 +6,7 @@
 package io.opentelemetry.extension.aws;
 
 import io.opentelemetry.context.propagation.TextMapPropagator;
+import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigurablePropagatorProvider;
 
 /**
@@ -14,7 +15,7 @@ import io.opentelemetry.sdk.autoconfigure.spi.ConfigurablePropagatorProvider;
  */
 public final class AwsConfigurablePropagator implements ConfigurablePropagatorProvider {
   @Override
-  public TextMapPropagator getPropagator() {
+  public TextMapPropagator getPropagator(ConfigProperties config) {
     return AwsXrayPropagator.getInstance();
   }
