+++ sdk-extensions/autoconfigure/build.gradle.kts
@@ -96,6 +96,7 @@ tasks {
     environment("OTEL_EXPORTER_OTLP_HEADERS", "cat=meow,dog=bark")
     environment("OTEL_EXPORTER_OTLP_TIMEOUT", "5000")
     environment("OTEL_SPAN_ATTRIBUTE_COUNT_LIMIT", "2")
+    environment("OTEL_TEST_CONFIGURED", "true")
   }
 
   val testInitializeRegistersGlobal by existing(Test::class) {
