+++ sdk-extensions/autoconfigure/src/main/java/io/opentelemetry/sdk/autoconfigure/OpenTelemetrySdkAutoConfiguration.java
@@ -8,6 +8,7 @@ package io.opentelemetry.sdk.autoconfigure;
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.context.propagation.ContextPropagators;
 import io.opentelemetry.sdk.OpenTelemetrySdk;
+import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.SdkMeterProviderConfigurer;
 import io.opentelemetry.sdk.metrics.SdkMeterProvider;
 import io.opentelemetry.sdk.metrics.SdkMeterProviderBuilder;
@@ -79,7 +80,7 @@ public final class OpenTelemetrySdkAutoConfiguration {
 
     for (SdkMeterProviderConfigurer configurer :
         ServiceLoader.load(SdkMeterProviderConfigurer.class)) {
-      configurer.configure(meterProviderBuilder);
+      configurer.configure(meterProviderBuilder, config);
     }
 
     SdkMeterProvider meterProvider = meterProviderBuilder.buildAndRegisterGlobal();
