+++ sdk-extensions/autoconfigure/src/main/java/io/opentelemetry/sdk/autoconfigure/PropagatorConfiguration.java
@@ -9,7 +9,9 @@ import io.opentelemetry.api.baggage.propagation.W3CBaggagePropagator;
 import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator;
 import io.opentelemetry.context.propagation.ContextPropagators;
 import io.opentelemetry.context.propagation.TextMapPropagator;
+import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigurablePropagatorProvider;
+import io.opentelemetry.sdk.autoconfigure.spi.ConfigurationException;
 import java.util.Arrays;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -28,7 +30,8 @@ final class PropagatorConfiguration {
             .collect(
                 Collectors.toMap(
                     ConfigurablePropagatorProvider::getName,
-                    ConfigurablePropagatorProvider::getPropagator));
+                    configurablePropagatorProvider ->
+                        configurablePropagatorProvider.getPropagator(config)));
 
     Set<TextMapPropagator> propagators = new LinkedHashSet<>();
     List<String> requestedPropagators = config.getCommaSeparatedValues("otel.propagators");
