+++ sdk-extensions/autoconfigure/src/main/java/io/opentelemetry/sdk/autoconfigure/spi/SdkMeterProviderConfigurer.java
@@ -12,10 +12,8 @@ import io.opentelemetry.sdk.metrics.SdkMeterProviderBuilder;
  * {@link SdkMeterProviderBuilder} during initialization. When using auto-configuration, you should
  * prefer to use system properties or environment variables for configuration, but this may be
  * useful to register components that are not part of the SDK such as registering views.
- *
- * @since 1.1.0
  */
 public interface SdkMeterProviderConfigurer {
   /** Configures the {@link SdkMeterProviderBuilder}. */
-  void configure(SdkMeterProviderBuilder meterProviderBuilder);
+  void configure(SdkMeterProviderBuilder meterProviderBuilder, ConfigProperties config);
 }
