+++ sdk-extensions/autoconfigure/src/main/java/io/opentelemetry/sdk/autoconfigure/spi/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/** Java SPI (Service Provider Interface) for implementing extensions to SDK autoconfiguration. */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.autoconfigure.spi;
+
+import javax.annotation.ParametersAreNonnullByDefault;
