+++ sdk-extensions/autoconfigure/src/testFullConfig/java/io/opentelemetry/sdk/autoconfigure/TestTracerProviderConfigurer.java
@@ -6,6 +6,7 @@
 package io.opentelemetry.sdk.autoconfigure;
 
 import io.opentelemetry.context.Context;
+import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.SdkTracerProviderConfigurer;
 import io.opentelemetry.sdk.trace.ReadWriteSpan;
 import io.opentelemetry.sdk.trace.ReadableSpan;
@@ -14,12 +15,12 @@ import io.opentelemetry.sdk.trace.SpanProcessor;
 
 public class TestTracerProviderConfigurer implements SdkTracerProviderConfigurer {
   @Override
-  public void configure(SdkTracerProviderBuilder tracerProvider) {
+  public void configure(SdkTracerProviderBuilder tracerProvider, ConfigProperties config) {
     tracerProvider.addSpanProcessor(
         new SpanProcessor() {
           @Override
           public void onStart(Context parentContext, ReadWriteSpan span) {
-            span.setAttribute("configured", true);
+            span.setAttribute("configured", config.getBoolean("otel.test.configured"));
           }
 
           @Override
