+++ sdk-extensions/resources/src/main/java/io/opentelemetry/sdk/extension/resources/HostResourceProvider.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.sdk.extension.resources;
 
-import io.opentelemetry.sdk.autoconfigure.ConfigProperties;
+import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider;
 import io.opentelemetry.sdk.resources.Resource;
